modularize player logic
